Logging: Moved level mapipulation to CLI and used caf-logging
